+++ 4. Fourth homework/bot/go.mod
@@ -1,0 +1,3 @@
+module bot
+
+go 1.17
