+++ 4. Fourth homework/bot/models.go
@@ -1,0 +1,29 @@
+package main
+
+type JsonData struct {
+	Name string `json:"name"`
+	Html string `json:"html_url"`
+}
+
+type Update struct {
+	UpdateId int     `json:"update_id"`
+	Message  Message `json:"message"`
+}
+
+type Message struct {
+	Chat Chat `json:"chat"`
+	Text string
+}
+
+type Chat struct {
+	ChatId int `json:"chat"`
+}
+
+type RestResponse struct {
+	Result []Update `json:"result"`
+}
+
+type BotMessage struct {
+	ChatId int    `json:"chat_id"`
+	Text   string `json:"text"`
+}
